# backend/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Code Review Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputFeedback } from './feedbackInputFeedback';

export interface FeedbackInput {
  suggestionId: number;
  feedback: FeedbackInputFeedback;
  note?: string;
}