# backend/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Code Review Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  title: string;
  language: string;
  status: string;
  suggestionCount: number;
  acceptedCount: number;
  rejectedCount: number;
  createdAt: Date;
}