# backend/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Code Review Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Suggestion {
  id: number;
  reviewId: number;
  category: string;
  severity: string;
  /** @nullable */
  lineStart?: number | null;
  /** @nullable */
  lineEnd?: number | null;
  message: string;
  /** @nullable */
  suggestedCode?: string | null;
  explanation: string;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  feedbackNote?: string | null;
  createdAt: Date;
}